folder tree adjusted, ajax functions moved to another file
