+++ src/components/UserView/UserView.js
@@ -3,7 +3,7 @@ import React from 'react';
 import { connect } from 'react-redux'
 import { Grid, Typography, Button } from '@material-ui/core';
 
-import { requestUserDelete } from "../../store/reducers/users"
+import { requestUserDelete } from "../../services"
 import { Link } from 'react-router-dom';
